Update account's amount when delete expense
