+++ src/expenses/controllers/expenses.controller.ts
@@ -4,13 +4,12 @@ import {
   UseGuards,
   Request,
   Put,
-  Delete,
   Get,
   Param,
 } from '@nestjs/common';
 import { JwtAuthGuard } from '../../auth/guards/jwt-auth.guard';
 import { ExpensesService } from '../services/expenses.service';
-import { DeleteExpenseDto, UpdateExpenseDto } from '../expenses.dto';
+import { UpdateExpenseDto } from '../expenses.dto';
 
 @UseGuards(JwtAuthGuard)
 @Controller('expenses')
@@ -20,12 +19,6 @@ export class ExpensesController {
   @Put('/multiple')
   updateMultipleExpense(@Body() payload: UpdateExpenseDto[]) {
     return this.expensesService.updateMultipleExpenses(payload);
-  }
-
-  @Delete()
-  removeExpense(@Body() payload: DeleteExpenseDto, @Request() req) {
-    const userId = req.user.sub;
-    return this.expensesService.removeExpense({ payload, userId });
   }
 
   @Get(':accountId/:month/:year')
